spark/requirements.txt
py4j==0.10.9.7
pyspark==3.4.1
python-dotenv==1.0.0
